+++ misc_components/locomotor-auto-config.js
